chore: rename addPackageMeta -> addBundleMeta
