+++ lib/add-bundle-meta.js
